restructure folder & routes
