+++ src/components/AddReview/SalaryReviewForm.jsx
@@ -5,13 +5,13 @@ import { Input, Switch, Form, Button, Icon, AutoComplete } from 'antd';
 import styled from 'styled-components';
 import { connect } from 'react-redux';
 import { withRouter } from 'react-router-dom';
-import { mobilePortrait, tabletPortrait } from '../styles/theme.styles';
-import { jobCategories } from '../utils/data';
-import currencies from '../utils/currencies';
-import Select from '../utils/select';
-import AutoCompleteComponent from '../utils/autocomplete';
-import { addSalaryReview } from '../state/actions/reviews';
-import openNotification from '../utils/openNotification';
+import { mobilePortrait, tabletPortrait } from '../../styles/theme.styles';
+import { jobCategories } from '../../utils/data';
+import currencies from '../../utils/currencies';
+import Select from '../../utils/select';
+import AutoCompleteComponent from '../../utils/autocomplete';
+import { addSalaryReview } from '../../state/actions/reviews';
+import openNotification from '../../utils/openNotification';
 
 const { TextArea } = Input;
 const { Option } = AutoComplete;
