+++ src/components/MyReviews/Company/SmallReviewCard.jsx
@@ -1,3 +1,5 @@
+/* eslint-disable jsx-a11y/no-noninteractive-element-to-interactive-role */
+/* eslint-disable jsx-a11y/click-events-have-key-events */
 /* eslint-disable react/jsx-one-expression-per-line */
 /* eslint-disable camelcase */
 import React from 'react';
@@ -47,7 +49,7 @@ export const SmallReviewCard = ({
     history.push(`/company-page/${company_id}`);
   };
   return (
-    <StyledCard onClick={() => history.push(`/reviews/${id}`)}>
+    <StyledCard onClick={() => history.push(`/reviews/company/${id}`)}>
       <div className="card-top">
         <h2 role="link" onClick={handleCompanyClick}>
           {name}
