--- src/components/ReviewDetails.jsx
@@ -1,15 +1,0 @@
-/* eslint-disable import/no-named-as-default */
-import React from 'react';
-import DetailedReviewCard from './MyReviews/CompanyReviews/DetailedReviewCard';
-
-const ReviewDetails = () => {
-  // eslint-disable-next-line no-shadow
-
-  return (
-    <div>
-      <DetailedReviewCard />
-    </div>
-  );
-};
-
-export default ReviewDetails;
