+++ src/components/Router/index.jsx
@@ -2,30 +2,32 @@
 import React from 'react';
 import { BrowserRouter, Switch, Route } from 'react-router-dom';
 import { decode } from 'jsonwebtoken';
+
+import DashboardLayout from '../Layout/DashboardLayout';
+
 import Home from '../../pages/Home/Home';
 import UserDashboard from '../../pages/UserDashboard/UserDashboard';
-import DashboardLayout from '../Layout/DashboardLayout';
-import ReviewDetails from '../ReviewDetails';
+import AddReview from '../../pages/AddReview/AddReview';
+import CompanyPage from '../../pages/Company/CompanyPage';
+import ManageReviews from '../../pages/MyReviews/ManageReviews';
+
 import DetailedSalaryReview from '../MyReviews/Salary/DetailedSalaryReviewCard';
-import AddReview from '../../pages/AddReview';
-// import ReviewList from '../ReviewList/ReviewList';
-import CompanyPage from '../../pages/CompanyPage';
+import DetailedCompanyReview from '../MyReviews/Company/DetailedReviewCard';
+import DetailedInterviewReview from '../MyReviews/Interview/DetailedReviewCard';
+
+import CompanyReview from '../Company/CompanyReview/CompanyReviewCardDetails';
+import InterviewReview from '../Company/InterviewReviews/InterviewReviewDetails';
+import SalaryReview from '../Company/SalaryReviews/SalaryReviewDetails';
+
 import store from '../../state/store';
 import { SetAuthenticated } from '../../state/actions/auth';
-import ManageReviews from '../../pages/ManageReviews';
 import {
   getCompanyReviews,
   getSalaryReviews,
   getInterviewReviews,
 } from '../../state/actions/reviews';
 import { getClosestCompanies } from '../../state/actions/closestCompanies';
-
 import { getCompanies } from '../../state/actions/companies';
-import DetailedCompanyReviewCard from '../MyReviews/CompanyReviews/DetailedReviewCard';
-import DetailedInterviewReviewCard from '../MyReviews/InterviewReviews/DetailedReviewCard';
-import CompanyReview from '../CompanyReview/CompanyReviewCardDetails';
-import InterviewReview from '../CompanyReviews/InterviewReviews/InterviewReviewDetails';
-import SalaryReview from '../CompanyReviews/SalaryReviews/SalaryReviewDetails';
 
 const start = async () => {
   const token = localStorage.getItem('token');
@@ -47,28 +49,23 @@ const AppRouter = () => {
       <Switch>
         <Route exact path="/" component={Home} />
         <DashboardLayout path="/dashboard" component={UserDashboard} />
-        <DashboardLayout path="/companies" component={UserDashboard} />
+        <DashboardLayout path="/add-review" component={AddReview} />
+        <DashboardLayout path="/company-page/:id" component={CompanyPage} />
         <DashboardLayout path="/reviews" exact component={ManageReviews} />
-        <DashboardLayout path="/interviews/:id" component={ReviewDetails} />
         <DashboardLayout
-          path="/salaries/:id"
-          component={DetailedSalaryReview}
+          path="/reviews/interview/:id"
+          component={DetailedInterviewReview}
         />
         <DashboardLayout
           path="/reviews/company/:id"
-          component={DetailedCompanyReviewCard}
+          component={DetailedCompanyReview}
         />
         <DashboardLayout
-          path="/reviews/interview/:id"
-          component={DetailedInterviewReviewCard}
+          path="/reviews/salary/:id"
+          component={DetailedSalaryReview}
         />
-        <DashboardLayout path="/add-review" component={AddReview} />
-        <DashboardLayout path="/company-page/:id" component={CompanyPage} />
         <DashboardLayout path="/companyReviews/:id" component={CompanyReview} />
-        <DashboardLayout
-          path="/interviewreviews/:id"
-          component={InterviewReview}
-        />
+        <DashboardLayout path="/interviewreviews/:id" component={InterviewReview} />
         <DashboardLayout
           path="/company/:companyId/salary/:id"
           component={SalaryReview}
