+++ src/pages/UserDashboard/UserDashboard.jsx
@@ -5,9 +5,9 @@ import React, { useEffect } from 'react';
 import axios from 'axios';
 import { connect } from 'react-redux';
 import styled from 'styled-components';
-import TopRatedList from '../../components/TopRated/TopRatedList';
-import ClosestLocationList from '../../components/UserLocationComp/ClosestLocationList';
-import JobTitleVisualization from '../../components/JobTitleVisualization';
+import TopRatedList from '../../components/UserDashboard/TopRated/TopRatedList';
+import ClosestLocationList from '../../components/UserDashboard/UserLocationComp/ClosestLocationList';
+import JobTitleVisualization from '../../components/UserDashboard/JobTitleVisualization';
 import { editProfile } from '../../state/actions/user';
 import { getLocation } from '../../utils/getLocation';
 import { LoginUser, SetAuthenticated } from '../../state/actions/auth';
